UHF-11015: Move getDecisionPdf() to bundle class
