+++ public/modules/custom/paatokset_ahjo_api/paatokset_ahjo_api.module
@@ -507,7 +507,7 @@ function _paatokset_ahjo_api_get_decision_variables(&$variables, CaseService $ca
   $decision = $caseService->getSelectedDecision();
 
   $variables['no_title_for_case'] = $caseService->checkIfNoCaseTitle();
-  $variables['decision_pdf'] = $caseService->getDecisionPdf();
+  $variables['decision_pdf'] = $decision?->getDecisionPdf();
   $variables['selectedDecision'] = $caseService->getSelectedDecision();
   $variables['policymaker_is_active'] = $caseService->decisionPmIsActive();
   $variables['decision_section'] = $decision?->getFormattedDecisionSection();
